Add: favor register form and remove ScrollView component
